Driver Function
Write a driver script that creates a binary
search tree from an array of random numbers < 100.
Confirm that the tree is balanced.
Print out all elements in pre, post, and in order.
Unbalance the tree by adding several numbers > 100.
Confirm that the tree is unbalanced by isBalanced.
Balance the tree by calling rebalance.